requirements.txt
coverage==4.3.4
Django==1.9.9
djangorestframework==3.4.1
flake8==3.3.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
